benchmark: move stuff to /internal to make it clear it's internal
