+++ benchmark/cmd/command/root.go
@@ -6,7 +6,7 @@ import (
 	"runtime"
 	"strings"
 
-	"github.com/pyroscope-io/pyroscope/benchmark/config"
+	"github.com/pyroscope-io/pyroscope/benchmark/internal/config"
 	"github.com/pyroscope-io/pyroscope/pkg/cli"
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
