+++ benchmark/internal/cireport/resources/image-report.gotpl
