+++ benchmark/internal/cireport/resources/pr.gotpl
